rewamped db/workflow.go to use pkg
